console: Refactor gateway overview to extract connect function
